changed i to type size_t

+++ src/my_alloc.c
@@ -128,7 +128,7 @@ static int get_size_class_index(size_t size)
 	int leftmost_ind = 0;
 	int iter_count = 0;
 
-	for(int i = 1 << 3; i > 0; i <<= 1){
+	for(size_t i = 1 << 3; i > 0; i <<= 1){
 		if((i & size) > 0){
 			leftmost_ind = iter_count;
 		}
